# openapi/agentteams-public.yaml
openapi: 3.1.0
info:
  title: AgentTeams Public API
  version: v1.0
  description: >-
    Stable public subset for Project and Task management. Internal control-plane,
    Matrix AppService, and Kubernetes APIs are intentionally excluded.
servers:
  - url: /
security:
  - BearerAuth: []
paths:
...
  /api/v1/tasks/{taskId}:
    get:
      operationId: getTask
      tags: [Tasks]
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Task
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/ServerError' }
...
  /api/v1/scheduled-tasks:
    get:
      operationId: listScheduledTasks
      tags: [ScheduledTasks]
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Scheduled task definitions
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/ScheduledTask' } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
    post:
      operationId: createScheduledTask
      tags: [ScheduledTasks]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateScheduledTaskRequest' }
      responses:
        '201':
          description: Created scheduled task
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ScheduledTask' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/ServerError' }
...
  /api/v1/scheduled-tasks/{scheduleId}:
    get:
      operationId: getScheduledTask
      tags: [ScheduledTasks]
      parameters:
        - $ref: '#/components/parameters/ScheduleId'
        - $ref: '#/components/parameters/OrganizationId'
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Scheduled task definition
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ScheduledTask' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }
  /api/v1/scheduled-tasks/{scheduleId}/pause:
    post:
      operationId: pauseScheduledTask
      tags: [ScheduledTasks]
      parameters:
        - $ref: '#/components/parameters/ScheduleId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ScheduleScopeRequest' }
      responses:
        '200':
          description: Paused scheduled task
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ScheduledTask' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/ServerError' }
  /api/v1/scheduled-tasks/{scheduleId}/resume:
    post:
      operationId: resumeScheduledTask
      tags: [ScheduledTasks]
      parameters:
        - $ref: '#/components/parameters/ScheduleId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ScheduleScopeRequest' }
      responses:
        '200':
          description: Resumed scheduled task
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ScheduledTask' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/ServerError' }
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 1, maxLength: 200 }
    TaskId:
      name: taskId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    ScheduleId:
      name: scheduleId
      in: path
      required: true
      schema: { type: string, format: uuid }
    OrganizationId:
      name: organizationId
      in: query
      required: true
      schema: { type: string, minLength: 1 }
    TenantId:
      name: tenantId
      in: query
      required: true
      schema: { type: string, minLength: 1 }
    ProjectId:
      name: projectId
      in: query
      required: false
      schema: { type: string, minLength: 1 }
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Forbidden:
      description: Resource is outside the caller scope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Conflict:
      description: Version or idempotency conflict
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    RateLimited:
      description: Retryable rate limit
      headers:
        Retry-After:
          schema: { type: integer, minimum: 0 }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    ServerError:
      description: Server or dependency failure
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
  schemas:
...
    ApiError:
      type: object
      required: [code, message, correlationId, details]
      properties:
        code: { type: string }
        message: { type: string }
        correlationId: { type: string }
        details:
          type: object
          additionalProperties: true
...
    Task:
      type: object
      required: [id, title, description, phase, priority, createdAt, updatedAt, version]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        description: { type: string }
        phase: { $ref: '#/components/schemas/TaskPhase' }
        priority: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        version: { type: integer, format: int64, minimum: 0 }
    CreateScheduledTaskRequest:
      type: object
      required: [name, organizationId, tenantId, projectId, cronExpression, timeZone, title]
      additionalProperties: false
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        organizationId: { type: string, minLength: 1 }
        tenantId: { type: string, minLength: 1 }
        projectId: { type: string, minLength: 1 }
        cronExpression: { type: string, minLength: 1 }
        timeZone: { type: string, minLength: 1 }
        title: { type: string, minLength: 1, maxLength: 500 }
        description: { type: string }
        spec: { type: object, additionalProperties: true }
        actor: { type: string }
        source: { type: string }
    ScheduleScopeRequest:
      type: object
      required: [organizationId, tenantId, projectId]
      additionalProperties: false
      properties:
        organizationId: { type: string, minLength: 1 }
        tenantId: { type: string, minLength: 1 }
        projectId: { type: string, minLength: 1 }
    ScheduledTask:
      type: object
      required: [id, name, organizationId, tenantId, projectId, cronExpression, timeZone, title, enabled, nextRunAt, version]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        organizationId: { type: string }
        tenantId: { type: string }
        projectId: { type: string }
        cronExpression: { type: string }
        timeZone: { type: string }
        title: { type: string }
        enabled: { type: boolean }
        nextRunAt: { type: string, format: date-time }
        lastRunAt: { type: [string, 'null'], format: date-time }
        lastTaskId: { type: [string, 'null'], format: uuid }
        version: { type: integer, format: int64, minimum: 0 }
...